# optimized matrix by 3x on cpu