used a direct fill approach and precomputed terms to reduce overhead, also about 1.3x faster on GPU

=== Device: cpu  B=100000 dtype=torch.float32 ===
Original     |   11.418 ms | orth err: 1.07e-06 | det err: 1.07e-06
Direct-fill  |    4.393 ms | orth err: 1.19e-06 | det err: 1.07e-06
Speedup (orig/fill):  2.60x
Max abs diff (orig vs fill): 5.960e-08
=== Device: cuda  B=100000 dtype=torch.float32 ===
Original     |    0.684 ms | orth err: 1.07e-06 | det err: 1.07e-06
Direct-fill  |    0.570 ms | orth err: 1.13e-06 | det err: 1.13e-06
Speedup (orig/fill):  1.20x
Max abs diff (orig vs fill): 5.960e-08


https://colab.research.google.com/drive/1DnAzeW-TlTmo3wXDMRqr3pcDtY7lzTdB?usp=sharing
here's the benchmark and validation script